news-agency/requirements.txt
langgraph
langgraph-cli[inmem]
langchain
langchain-openai
python-dotenv
httpx
beautifulsoup4
pydantic